pyproject.toml
[tool.poetry]
name = "feedback-tg-bot"
version = "0.1.0"
description = ""
authors = ["pavlo <iris398322@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.8.0"
sulguk = "^0.8.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
